refactor: Moved portfolio link to new footer component. Removed navbar component.
