--- gifts/src/NavigationBar/NavigationBar.css
@@ -1,21 +1,0 @@
-nav {
-  background-color: rgb(124, 124, 124, 0.7);
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
-
-nav ul {
-  list-style: none;
-  /* padding-left: 0.5rem; */
-  max-width: 10vw;
-  margin: 0;
-  padding: 0.5rem 0.5rem;
-}
-
-.divider {
-  color: black;
-}
-
-
